requirements.txt
fastapi
uvicorn
sentence-transformers
faiss-cpu     # or faiss-gpu if you're using GPU
numpy
pandas
streamlit
requests
